revisi unit testing

+++ tests/Unit/API/AuthTest.php
@@ -10,6 +10,11 @@ use App\Models\User;
 
 class AuthTest extends TestCase
 {
+    /**
+     * A basic unit test example.
+     *
+     * @return void
+     */
     public function test_register()
     {
         $response = $this->json('POST', 'api/v1/register', [
@@ -33,12 +38,13 @@ class AuthTest extends TestCase
 
     public function test_logout()
     {
-        auth()->attempt(['email' => 'testauth@example.com', 'password' => 'testauthpassword']);
+        if(!auth()->attempt(['email' => 'testauth@example.com', 'password' => 'testauthpassword'])){
+            return response(['message' => 'Failed']);
+        }
 
         $token = auth()->user()->createToken('LaravelAuthApp')->accessToken;
 
-        $response = $this->withHeaders(['Authorization' => 'Bearer'. $token])->json('POST', 'api/v1/logout');
-        
+        $response = $this->withHeaders(['Authorization' => 'Bearer '. $token])->json('POST', 'api/v1/logout');
         $response->assertStatus(200);
 
         User::where('email','testauth@example.com')->delete();
